+++ Models/SoundEffects/SoundEffect.cs
@@ -1,5 +1,4 @@
-﻿using static MMRando.Models.SoundEffects.SoundEffectAttributes;
-using static MMRando.Models.SoundEffects.SoundEffectTag;
+﻿using static MMRando.Models.SoundEffects.SoundEffectTag;
 
 namespace MMRando.Models.SoundEffects
 {
@@ -17,58 +16,80 @@ namespace MMRando.Models.SoundEffects
 
         [Replacable(0x00EABA46,0x00EABC3A, 0x00EACACA), Tags(Long)]
         CuccoMorning = 0x2013,
+
         [Replacable(0x00EABC4A, 0x00EACABE), Tags(Long)]
         WolfHowlEvening = 0x20AE, 
+
         [Replacable(0x00DFC776), Tags(Short, LowHpBeep), ReplacableByTags(Short)]
         DogBark = 0x20D8,
+
         [Tags(Long)]
         CowMoo = 0x20DF,
+
         [Replacable(0x00DFC7B6), Tags(Short, LowHpBeep), ReplacableByTags(Short)]
         DogBarkAngry = 0x2110,
 
         // System
 
-        [Replacable(0x00B97E2A), Tags(Short), ReplacableByTags(LowHpBeep)]
+        [Replacable(0x00B97E2A), Tags(Short, LowHpBeep), ReplacableByTags(LowHpBeep)]
         LowHealthBeep = 0x401B,
 
         // Voice
 
         [Tags(Short)]
         FierceDeityLinkAttack = 0x6000,
+
         [Tags(Short)]
         FierceDeityLinkJumpAttack = 0x6001,
+
         [Tags(Short)]
         ChildLinkAttack = 0x6020,
+
         [Tags(Short)]
         ChildLinkJumpAttack = 0x6021,
+
         [Tags(Short)]
         ChildLinkGrabLedge = 0x6023,
+
         [Tags(Short)]
         ChildLinkMountLedge = 0x6024,
+
         [Tags(Short, LowHpBeep)]
         GoronPunch = 0x60C0,
+
         [Tags(Short, LowHpBeep)]
         JimHeh = 0x6100,
+
         [Tags(Short, LowHpBeep)]
         BomberGiggle = 0x6101,
+
         [Tags(Short, LowHpBeep)]
         JimHuh = 0x6102,
+
         [Tags(Short, LowHpBeep)]
         BomberEhh = 0x6103,
+
         [Tags(Short, LowHpBeep)]
         ReceptionistMmHmm = 0x612A,
+
         [Tags(Short, LowHpBeep)]
         ReceptionistMmm = 0x612C,
+
         [Tags(Short, LowHpBeep)]
         MutohScoff = 0x614F,
+
         [ReplacableInMessage(0x696F, 10100, 10101, 10107, 10111, 10114, 10123, 10141, 10149, 10155, 10427, 10429, 10431, 10589), Tags(Short, LowHpBeep)]
         PostmanGreeting = 0x616F,
+
         [Tags(Short, LowHpBeep)]
         MikauBaybee = 0x6175,
+
         [Tags(Short, LowHpBeep)]
         MikauYay = 0x6176,
+
         [Tags(Long)]
         TingleChuckle = 0x617A,
+
         [ReplacableInMessage(0x6959, 10540, 10541, 10543, 10547, 10573, 10574), Tags(Long)]
         GuruGuruLalala = 0x6159,
     }
